SSP/requirements.txt
# pip install -r requirements.txt

pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
imbalanced-learn==0.12.2
python-dotenv>=1.0.0
pymysql>=1.1.0